Added SQLite services to the application
everything in database\SQLiteStudio\ is to have a GUI to help with creating tables and populating tables. Not needed in the final version

everything in \database is needed

the changes in SSL-Index\app-documents\playerAttributes.R, palyerStats.R, standings.R , teamOverview.R is to change column names to be the same as the SQLite column names.

           ex, everything with a space now has a _ like Minutes_Played
                 everything with a % is percentage, like Save_Percentage

what is in SSL-Index\app-documents\dataLoader.R is the change to SQLite queries

the change in app.R is adding the packages for SQLite